Split Cardano.Api.ProtocolParameters module to reduce compilation time
